+++ exercises/Samuel_Jones/ex02_oo_basics/src/Rectangle.cpp
@@ -4,10 +4,10 @@ Rectangle::Rectangle(int side1, int side2):m_side1(side1), m_side2(side2), Shape
 {
 }
 
-float Rectangle::getPerimeter(){
+double Rectangle::getPerimeter(){
     return m_side1*2 + m_side2*2;
 }
 
-float Rectangle::getArea(){
+double Rectangle::getArea(){
     return m_side1*m_side2;
 }
