+++ exercises/Samuel_Jones/ex02_oo_basics/src/Shape.h
@@ -11,8 +11,8 @@ public:
     Shape(std::string type);
     std::string getType();
     virtual ~Shape() = default;
-    virtual float getPerimeter();
-    virtual float getArea();
+    virtual double getPerimeter();
+    virtual double getArea();
 };
 
 #endif /*SHAPE_H_*/
