+++ exercises/Samuel_Jones/ex02_oo_basics/src/Square.cpp
@@ -4,10 +4,10 @@ Square::Square(int side1):m_side1(side1), Shape("Square")
 {
 }
 
-float Square::getPerimeter(){
+double Square::getPerimeter(){
     return m_side1*4;
 }
 
-float Square::getArea(){
+double Square::getArea(){
     return m_side1*m_side1;
 }
