+++ exercises/Samuel_Jones/ex02_oo_basics/src/Triangle.cpp
@@ -4,10 +4,10 @@ Triangle::Triangle(int base, int height):m_base(base), m_height(height), Shape("
 {
 }
 
-float Triangle::getPerimeter(){
+double Triangle::getPerimeter(){
     return m_base + 2*(sqrt((m_height*m_height)+((m_base*m_base)/4)));
 }
 
-float Triangle::getArea(){
+double Triangle::getArea(){
     return (m_base*m_height)/2;
 }
